Deletion Product

+++ GamifiedMarketingEJB/ejbModule/services/ProductService.java
@@ -6,6 +6,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.NonUniqueResultException;
@@ -14,6 +15,8 @@ import javax.persistence.PersistenceException;
 
 import entities.Admin;
 import entities.Product;
+import entities.Question;
+import entities.Submission;
 import exceptions.CredentialsException;
 
 @Stateless
@@ -21,15 +24,22 @@ public class ProductService {
 	@PersistenceContext(unitName = "GamifiedMarketingEJB")
 	private EntityManager em;
 
+
+	@EJB(name = "services/QuestionService") 
+	private QuestionService questionService;
+
+	@EJB(name = "services/SubmissionService") 
+	private SubmissionService submissionService;
+
 	public ProductService() {
 	}
-	
+
 	public Product findDailyProduct() throws NonUniqueResultException {
 		List<Product> aList = null;
 		LocalDate localDate = LocalDateTime.now().plus(Duration.ofHours(2)).toLocalDate();
 		Date date = Date.valueOf(localDate);
 		aList = em.createNamedQuery("Product.findDailyProduct", Product.class).setParameter(1, date)
-					.getResultList();
+				.getResultList();
 		if (aList.isEmpty())
 			return null;
 		else if (aList.size() == 1)
@@ -37,11 +47,11 @@ public class ProductService {
 		throw new NonUniqueResultException("More than one daily products detected");
 
 	}
-	
+
 	public List<Product> getAllProducts()  {
 		return em.createNamedQuery("Product.getAllProducts", Product.class).getResultList();
 	}
-	
+
 	public List<Product> findPreviousProduct()  {
 		List<Product> aList = null;
 		LocalDate localDate = LocalDateTime.now().plus(Duration.ofHours(2)).toLocalDate();
@@ -52,28 +62,36 @@ public class ProductService {
 		else
 			return aList;
 	}
-	
+
 	public Product getProductById(int productId) {
 		return em.createNamedQuery("Product.getProductById", Product.class).setParameter("productId", productId).getSingleResult();
 	}
-	
+
 	public Product addProduct(String name, byte[] image, Date date) {
 		Product product = new Product(name, image, date);
 		em.persist(product);
 		em.flush();
 		return product;
 	}
-	
+
 	public void deleteProduct(int productId) {
-        Product product = em.find(Product.class, productId);
-        
+		Product product = em.find(Product.class, productId);
 
-        if (product == null) {
-            throw new IllegalArgumentException(String.format("Product with ID = %d does not exist!", productId));
-        }
 
-        em.remove(product);
-        em.flush();
-    }
-	
+		List<Submission> submissions = submissionService.findByProduct(productId);
+
+
+		List<Question> questions = questionService.getQuestions(productId);
+
+		if (product == null || submissions == null) {
+			throw new IllegalArgumentException(String.format("Product with ID = %d does not exist!", productId));
+		}
+		for(Submission s: submissions)
+			{em.remove(s);}
+		for(Question q: questions)
+			{em.remove(q);}
+		em.remove(product);
+		em.flush();
+	}
+
 }
